src/requirements.txt
kaggle
pandas
pascal-voc
scikit-image
ultralytics